Restructure file hierarchy, with one file per route function. Add event routes & function names.
