+++ server/api/functions/announcements/deleteAnnouncement.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for DELETE: /announcements/delete/{postId}
+export function deleteAnnouncement(req:Request, res:Response) {
+    res.send("deleted")
+}
