+++ server/api/functions/announcements/getAnnouncements.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for GET: /announcements/latest/{offset}/{entryCount}
+export function getAnnouncements(req:Request, res:Response) {
+    res.send("got the announcements!");
+}
