+++ server/api/functions/announcements/index.ts
@@ -1,0 +1,8 @@
+// Barrel for announcement functions
+
+import { deleteAnnouncement } from "./deleteAnnouncement";
+import { getAnnouncements } from "./getAnnouncements";
+import { getAnnouncementCount } from "./getAnnouncementCount";
+import { upsertAnnouncement } from "./upsertAnnouncement";
+
+export { deleteAnnouncement, getAnnouncements, getAnnouncementCount, upsertAnnouncement };
