+++ server/api/functions/announcements/upsertAnnouncement.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for POST: /announcements/post/{postId}
+export function upsertAnnouncement(req:Request, res:Response) { 
+    res.send("posted")
+}
