+++ server/api/functions/events/deleteEvent.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for DELETE: /events/delete/{eventId}
+export function deleteEvent(req:Request, res:Response) {
+    res.send("deletedEvent")
+}
