+++ server/api/functions/events/getEventCount.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for GET: /events/count/{startDate}/{endDate}
+export function getEventCount(req:Request, res:Response) {
+    res.send("1");
+}
