+++ server/api/functions/events/getEvents.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for GET: /events/all/{entryCount}/{offset}
+export function getEvents(req:Request, res:Response) {
+    res.send("got the events!");
+}
