+++ server/api/functions/events/index.ts
@@ -1,0 +1,6 @@
+import { getEvents } from "./getEvents";
+import { getEventCount } from "./getEventCount";
+import { deleteEvent } from "./deleteEvent";
+import { upsertEvent } from "./upsertEvent";
+
+export {getEvents, getEventCount, deleteEvent, upsertEvent};
