+++ server/api/functions/events/upsertEvent.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+
+// Used for POST: /events/post/{eventId}
+export function upsertEvent(req:Request, res:Response) { 
+    res.send("event posted")
+}
